Move 'triggerable' to private folder
