+++ addon/index.js
@@ -4,7 +4,7 @@ import { clickable }   from './-private/properties/clickable';        export { c
 import { clickOnText } from './-private/properties/click-on-text';    export { clickOnText };
 import { fillable }    from './-private/properties/fillable';         export { fillable }; export const selectable = fillable;
 import { visitable }   from './actions/visitable';        export { visitable };
-import { triggerable } from './actions/triggerable';      export { triggerable };
+import { triggerable } from './-private/properties/triggerable';      export { triggerable };
 import { contains }    from './predicates/contains';      export { contains };
 import { hasClass }    from './predicates/has-class';     export { hasClass };
 import { isHidden }    from './predicates/is-hidden';     export { isHidden };
